Fixes Bug: 250
A previously existing output file produced by a processor is saved.
When a new input file is processed by a processor, it is now checked whether the
output that is produced is good. If so, it is retained, else the previously
saved file (from a previous run) is copied to the output location.