fixes bug [256] -
The process of converting a stocklist to a price-file and finding the missing isbns
is now performed within a function processfile()
Another function added to this script is to check if a book in the newarrivals file is carried
by the other stocklists. If this is true then these books must be removed from newarrivals file.

Another requirement that will be satisfied by the script is to process the newarrivals only after all the
other stocklists have been processed.

The missing_isbns table in the reference database will also be populated