Significant. Import only changed catalog files.
importbooks has been changed to rely on the contents of a new table,
ek_import_state, that holds information about when data files are loaded.
On any import, this table is consulted to see if the data file is newer
than the date of previous import. If yes, the file is imported. If no, the
file is not imported.

This commit also has changes to control when the book_promo table
should be copied to ekkitab_books from the reference database. If there are
changes, then the table is copied. Else not. The same holds for the release
to the production server.